feat: add `ServiceError`
