+++ infra/errors.js
@@ -1,18 +1,37 @@
 export class InternalServerError extends Error {
-  constructor({ cause }) {
+  constructor({ cause, statusCode }) {
     super("Um erro interno não esperado aconteceu.", {
       cause,
     });
     this.name = "InternalServerError";
     this.action = "Entre em contato com o suporte.";
-    this.status_code = 500;
+    this.statusCode = statusCode || 500;
   }
   toJSON() {
     return {
       name: this.name,
       message: this.message,
       action: this.action,
-      status_code: this.status_code,
+      statusCode: this.statusCode,
+    };
+  }
+}
+
+export class ServiceError extends Error {
+  constructor({ cause, message }) {
+    super(message || "Serviço indisponível no momento.", {
+      cause,
+    });
+    this.name = "InternalServerError";
+    this.action = "Verifique se o serviço está disponível.";
+    this.statusCode = 503;
+  }
+  toJSON() {
+    return {
+      name: this.name,
+      message: this.message,
+      action: this.action,
+      statusCode: this.statusCode,
     };
   }
 }
@@ -23,14 +42,14 @@ export class MethodNotAllowedError extends Error {
     this.name = "MethodNotAllowedError";
     this.message = "Método não permitido para esse endpoint.";
     this.action = "Verifique se o método HTTP enviado é válido para esse endpoint.";
-    this.status_code = 405;
+    this.statusCode = 405;
   }
   toJSON() {
     return {
       name: this.name,
       message: this.message,
       action: this.action,
-      status_code: 405,
+      statusCode: 405,
     };
   }
 }
